Integrate pattern intersection regressions into existing tests
